Switch audio from 8-bit PCM to 16-bit PCM which seems to have had a DRASTIC improvement for audio quality

+++ projectrovercommunications/src/main/java/xyz/philiprodriguez/projectrovercommunications/ProjectRoverClient.java
@@ -175,7 +175,7 @@ public class ProjectRoverClient {
         senderThread.enqueueStrict(armPositionMessage);
     }
 
-    public void doEnqueueAudioFrame(byte[] pcmValues) {
+    public void doEnqueueAudioFrame(short[] pcmValues) {
         senderThread.enqueueDroppable(new PCMFrameMessage(System.currentTimeMillis(), pcmValues));
     }
